Show book index invalid even for integers larger than Integer.MAX

+++ src/main/java/seedu/address/logic/commands/RenewCommand.java
@@ -64,7 +64,7 @@ public class RenewCommand extends ReversibleCommand {
             throw new CommandException(MESSAGE_INVALID_BOOK_DISPLAYED_INDEX);
         }
 
-        Book bookToBeRenewed = lastShownBorrowerBooksList.get(index.getZeroBased());
+        Book bookToBeRenewed = lastShownBorrowerBooksList.get((int) index.getZeroBased());
         if (!bookToBeRenewed.isCurrentlyLoanedOut()) {
             throw new CommandException(String.format(MESSAGE_BOOK_NOT_ON_LOAN, bookToBeRenewed));
         }
